Back of Card and Printer Settings
This commit adds the start of the back of the card as well as makes some additional changes to the print formatting in the app.css.

The back has only been partially implemented.  I need to create a new background image (or get more clever with CSS), and then create 2 new angle box CSS classes for the mirror side on the back. Those changes are only in the CSS file though, so the back is now good for use.

You no longer need to check the "hide input" button in order to print, as the print @media tag will hide everything except the card.  Because this worked so well, I moved the input area over to float: left so that the card would be right beside. I need to make a few more changes to the input area to allow it to scroll so the user can stay focused on the card output.