Prototype Coding Results
This commit has a bunch of prototype changes. First, I tried a new approach with how the cube is displayed in order to start utilizing a higher DPI image to make printing look nicer. It also changes how items on the back look, and adds the folded corner to the back.