Rework Front to Not Use Background Image
This commit reworks the front of the card to no longer use the
background image, and instead create everything from scratch.
Unfortunately, you lose some of the nice touches, like the corners being
cut out of the image.  That said, it should print cleaner and look
nicer. It also opens the door for allowing user-defined colors
everywhere.