Minor Cleanup Tweak
Just a simple change to make life a little easier.

Some of these changes were found because of a massive failure. I tried
to add the bootstrap AngularJS library, but the requirement of the
bootstrap CSS and its media overwrites kept me from being able to use
it. So for now, I will have to create my own directive unfortunately.