Adding Save To PC
This commit adds a new button that when clicked will create a json file
for the user to save.  The new file uses the name of the unit to create
a file like "Travis Young.json".  The hope is that these files can be
later uploaded easily by users to restore their content.