Cards as Templates
This commit moves the card fronts and backs into templates.
Essentially, they are each their own view with a shared model and
controller.  This will allow me to make a page version where users can
print 4 cards at a time and reuse this code.  Kinda neat!

Unfortunately, there is still a bug with the size when printing, at
least when printing to PDF in Chrome.  I believe the issue has to do
with Chrome itself, but I'm uncertain how I can go about fixing it.