Color Selection Initial Version
So much for the background-picker branch. Due to needing to use :after
for the cut corner stuff, I cannot change the background color.  That
said, every other color can change, and this initial version contains
those changes in the form of a huge input block using an open source
color picker.