ngDialog Integration and Initial Add Skill
This commit introduces the ngDialog and Add Skill dialog from that.
Unfortunately, the dialog isn't performing as expected.  It seems that
the skills array is being updated, but I cannot figure out how to get
the result properly applied in the context of a button from the add
dialog.  I think I may have to resort to relying on the promise and
adding it there instead.